End of march 29

+++ db/migrate/20200329160918_create_provinces.rb
@@ -2,9 +2,9 @@ class CreateProvinces < ActiveRecord::Migration[6.0]
   def change
     create_table :provinces do |t|
       t.string :name
-      t.float :GST_rate
-      t.float :PST_rate
-      t.float :HST_rate
+      t.integer :GST_rate
+      t.integer :PST_rate
+      t.integer :HST_rate
 
       t.timestamps
     end
